회원용 게시판 추가 외 7건
필터 짤림 현상 수정
웹사이트 좌우폭 고정
댓글 css 수정
애니메이션 폭죽 효과 추가
스크래핑 모듈 cmd 변수 입력 기능 추가
스크래핑 코드 리팩터링
키워드 검색 코드 수정

+++ src/js/update_titles.js
@@ -2,12 +2,13 @@
 Object.defineProperty(exports, '__esModule', { value: true })
 const { client, connectCollection } = require('./mongo')
 const crawling_1 = require('./crawling')
-async function main() {
+async function main(URI, reset) {
   try {
     const games = await connectCollection('games')
-    const gameList = await (0, crawling_1.default)()
-    // Reset
-    await games.deleteMany({})
+    const gameList = await (0, crawling_1.default)(URI)
+    if (reset === 'y') {
+      await games.deleteMany({})
+    }
     await games.insertMany(gameList)
     await client.close()
     console.log('Database updated!!')
@@ -15,4 +16,4 @@ async function main() {
     console.log(error)
   }
 }
-main()
+main(process.argv[2], process.argv[3])
